+++ app/helpers/application_helper.rb
@@ -1,17 +1,3 @@
 # frozen_string_literal: true
 
-module ApplicationHelper
-  def current_organization_admin
-    current_referencee.instance_of?(OrganizationMember) ? current_referencee : nil
-  end
-
-  def current_civil_servant
-    current_referencee.instance_of?(CivilServant) ? current_referencee : nil
-  end
-
-  private
-
-  def current_referencee
-    current_user&.referencee
-  end
-end
+module ApplicationHelper; end
