+++ app/javascript/stylessheets/organizations/organization_members.scss
@@ -4,7 +4,7 @@
   width: 100px;
   height: 100px;
   background-image: linear-gradient($light-gray, $gray);
-  //background: $gray;
+  background-color: $gray;
   border-radius: 50%;
   color: $white;
   font-size: 2.5rem;
