+++ spec/helpers/organization_members_helper_spec.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe OrganizationMembersHelper, type: :helper do
+  describe '#initials' do
+    subject { helper.initials(organization_member) }
+
+    let(:organization_member) { build_stubbed(:organization_member, first_name: 'Peter', last_name: 'warlowsk') }
+
+    it { is_expected.to eq 'PW' }
+  end
+end
